# Relax ownership requirements on Cached<K,V>