--- a/src/stores/timed.rs
+++ b/src/stores/timed.rs
@@ -99,7 +99,11 @@ impl<K: Hash + Eq, V> TimedCache<K, V> {
 }
 
 impl<K: Hash + Eq, V> Cached<K, V> for TimedCache<K, V> {
-    fn cache_get(&mut self, key: &K) -> Option<&V> {
+    fn cache_get<Q>(&mut self, key: &Q) -> Option<&V>
+    where
+        K: std::borrow::Borrow<Q>,
+        Q: std::hash::Hash + Eq + ?Sized,
+    {
         let status = {
             let mut val = self.store.get_mut(key);
             if let Some(&mut (instant, _)) = val.as_mut() {
@@ -132,7 +136,11 @@ impl<K: Hash + Eq, V> Cached<K, V> for TimedCache<K, V> {
         }
     }
 
-    fn cache_get_mut(&mut self, key: &K) -> Option<&mut V> {
+    fn cache_get_mut<Q>(&mut self, key: &Q) -> Option<&mut V>
+    where
+        K: std::borrow::Borrow<Q>,
+        Q: std::hash::Hash + Eq + ?Sized,
+    {
         let status = {
             let mut val = self.store.get_mut(key);
             if let Some(&mut (instant, _)) = val.as_mut() {
@@ -198,7 +206,11 @@ impl<K: Hash + Eq, V> Cached<K, V> for TimedCache<K, V> {
             }
         })
     }
-    fn cache_remove(&mut self, k: &K) -> Option<V> {
+    fn cache_remove<Q>(&mut self, k: &Q) -> Option<V>
+    where
+        K: std::borrow::Borrow<Q>,
+        Q: std::hash::Hash + Eq + ?Sized,
+    {
         self.store.remove(k).and_then(|(instant, v)| {
             if instant.elapsed().as_secs() < self.seconds {
                 Some(v)